backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
uuid
psycopg2-binary
flask-cors
